lang

+++ database/migrations/2020_06_12_094251_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             $table->bigIncrements('id');
             $table->integer('num_of_orders')->default('0');
             $table->enum('status', array('فعال','منتهي', 'ملغي','منتظر'));
-            $table->integer('total')->default(0);
+            $table->double('total')->default(0);
             $table->double('tax_price')->default('0');
             $table->unsignedBigInteger('client_id')->nullable();
             $table->timestamps();
